% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{chop_url}
\alias{chop_url}
\title{Chop off URL subdomains and directories}
\usage{
chop_url(url, n_dirs = 1, n_subdomains = 1, show_regex = FALSE)
}
\arguments{
\item{url}{Vector of URLs to process}

\item{n_dirs}{Number of directories (e.x., lukedav.is/_blog_) to keep.
Pass `NULL` (the default) to keep all and 0 to remove all.}

\item{n_subdomains}{Number of subdomains (e.x., _blog_.lukedav.is) to keep.
#'   Pass `NULL` (the default) to keep all and 0 to remove all.}

\item{show_regex}{Should the generated regex be printed? This can be useful
if you want to repeat the same "chopping" in another system that supports
regex, like a SQL query.}
}
\value{
A vector of URLs the same length as the input with directories
  and / or subdomains truncated.
}
\description{
Chop off URL subdomains and directories
}
\details{
For more general URL extraction tasks check out the urltools package.
  This function should _not_ be used to verify if an input is a valid URL;
  Its rules for TLDs (e.x., ".com", ".org") are extremely generous and it
  isn't intended to validate the URL format more broadly.
}
