% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract_url_param}
\alias{extract_url_param}
\title{Extract parameters from a URL}
\usage{
extract_url_param(url, param)
}
\arguments{
\item{url}{Character vector of URLs}

\item{param}{The name of the parameter to extract as a string}
}
\description{
Extract parameters from a URL
}
