% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fiscal_quarter}
\alias{fiscal_quarter}
\title{Quarter within fiscal year}
\usage{
fiscal_quarter(date, start_month = 7, start_day = 1)
}
\arguments{
\item{date}{A vector of dates}
}
\value{
An integer vector the same length as `date` giving the quarter in the
fiscal year that each date falls in.
}
\description{
Quarter within fiscal year
}
\details{
Assumes that quarters are each exactly three months and start on
  the same day of the month.
  If you want both fiscal year and quarter you can use `fiscal_year`
  together with `fiscal_quarter`.
}
