% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fiscal_year}
\alias{fiscal_year}
\title{Fiscal year of a date}
\usage{
fiscal_year(date, start_month = 7, start_day = 1)
}
\arguments{
\item{date}{A vector of dates}

\item{start_month}{Integer giving the first month of the fiscal year}

\item{start_day}{Integer giving the first day of the fiscal year}
}
\value{
An integer vector the same length as `date` giving the fiscal year
for each respective date.
}
\description{
Many organizations use "fiscal years" as their primary time period for
calculating and reporting metrics. This function lets you specify the month
and day of the fiscal year and calculate the fiscal year of each date in a
vector.
}
