% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fiscal_year_day}
\alias{fiscal_year_day}
\title{Number of days since start of fiscal year}
\usage{
fiscal_year_day(date, start_month = 7, start_day = 1)
}
\arguments{
\item{date}{A vector of dates}

\item{start_month}{Integer giving the first month of the fiscal year}

\item{start_day}{Integer giving the first day of the fiscal year}
}
\value{
A numeric vector the same length as `dates` giving
}
\description{
Number of days since start of fiscal year
}
