% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fiscal_year_quarter}
\alias{fiscal_year_quarter}
\title{Fiscal year and quarter}
\usage{
fiscal_year_quarter(date, sep = "-")
}
\arguments{
\item{date}{A vector of dates}

\item{sep}{Character separator passed on to `paste`}
}
\value{
A character vector the same length as `date` giving the quarter and
  year within which each date falls.
}
\description{
Fiscal year and quarter
}
\details{
Since this is not a standard date format the return result is a
  character vector. By default the year and quarter are separated by a "-",
  e.x., "2020-Q2".
}
