% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{get_or_set_credentials}
\alias{get_or_set_credentials}
\title{Retrieve or set keychain credentials}
\usage{
get_or_set_credentials(service, force_ask = FALSE)
}
\value{
A list with two elements: "username" and "password"
}
\description{
This function queries the system keychain for credentials for a given service
and gives user prompts (in RStudio) if no existing credentials are found.
}
