% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%between\%}
\alias{\%between\%}
\title{Check if elements are between two values}
\usage{
x \%between\% y
}
\value{
A logical vector indicating whether each element of the left hand
side was between the min and max of the right hand side.
}
\description{
This is a convenience syntax for the common pattern of `x >= y & x <= z`.
The right hand side can be supplied as a simple vector: e.x., `x %between% c(y, z)`.
The min and max of the right hand side vector are used for the comparison, so
the order of the elements in the vector doesn't matter. This also means you
could use the function to succinctly find whether or not an element falls
within the range of another vector.
}
\details{
Both sides of the comparison are inclusive. For example,
`1 %between% c(1, 3)` evaluates to `TRUE`, and so does `3 %between% c(1, 3)`.
}
