% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%like\%}
\alias{\%like\%}
\title{Match strings with SQL-like syntax}
\usage{
string \%like\% pattern
}
\value{
A logical vector indicating whether each element had a match.
}
\description{
This is a convenience wrapper around `stringr::str_detect`.
}
\details{
Inspired by `data.table::like`, which uses `base::grepl` instead.
  Because this function uses `stringr::str_detect` the second argument
  (the pattern) can also be a vector.

  See the help page for [stringr::str_detect()] for more details.
}
