% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{holiday_dates}
\alias{holiday_dates}
\title{Get dates of holidays}
\usage{
holiday_dates(year, country = "US")
}
\arguments{
\item{year}{A vector of one or more years for which to list holiday dates}

\item{country}{Two-letter country code}
}
\value{
Named character vector of holiday dates as "YYYY-MM-DD".
The vector's names will be the concatenation of the country code, the
holiday name, and the year.
}
\description{
Gives a vector of each holiday date
}
