% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{list_to_sql}
\alias{list_to_sql}
\title{Convert an R vector to a SQL list}
\usage{
list_to_sql(x, con)
}
\arguments{
\item{x}{Vector or list to be converted}

\item{con}{Database connection object}
}
\value{

}
\description{
Convert an R vector to a SQL list
}
