% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pmedian}
\alias{pmedian}
\title{Compute parallel medians}
\usage{
pmedian(..., na.rm = TRUE)
}
\description{
Computes the element-wise median of multiple vectors.
}
\details{
The motivation behind this function is the base R functions `pmin`
  and `pmax`.

A more general way to approach this type of problem is with
  `[purrr::lift()]`.
}
