% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rolling}
\alias{rolling}
\title{Rolling window computations}
\usage{
rolling(fun, x, n, default = NA, forward = FALSE, fun_value = NULL, ...)
}
\arguments{
\item{fun}{Bare, unquoted function to be applied over the window}

\item{x}{Vector to which \code{fun} will be applied}

\item{n}{Size of the window}

\item{default}{Default value used for non-existent values on either end of
\code{x}}

\item{forward}{A logical value indicating whether the window should go
forwards from each value or backwards (the default)}

\item{fun_value}{A generalized vector to serve as a template for the return
value from \code{fun}. If \code{NULL} (the default) it is assumed that
the return value is the same type as \code{x}. \code{fun_value} is
simply passed to \code{vapply}, so it uses the same format.}
}
\value{
A vector with the same length as \code{x}.
}
\description{
Computes the value of a summary function over a rolling window of nearby
values in the vector.
}
\examples{
  rolling(mean, rnorm(50), n = 10, default = 0.5)
  rolling(paste, 1:10, n = 3, forward = TRUE, fun_value = character, collapse = ",")
}
