% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{run_time}
\alias{run_time}
\title{Print the time taken to run an expression}
\usage{
run_time(expr, message = NULL, digits = 2)
}
\arguments{
\item{expr}{An expression to be evaluated.}

\item{message}{A string describing what you're doing. This will be appended
to the start of the printed message. For example, if you set
`message = "run things"` the output will look like
`Time taken to run things: 5.15 seconds`.}
}
\value{
The result of evaluating `expr` in the environment from which
`run_time` was called.
}
\description{
Sometimes you just want to know how long something took. There are many
more elaborate solutions for more formal benchmarking (like microbenchmark),
but for some applications you don't want to repeat an action or do formal
benchmarking. This function simply evaluates the expression in `expr`
and outputs the time taken as a `message`.
}
\seealso{
`\link{system.time}` base R solution with more basic output.
}
