% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale_zero_to_one}
\alias{scale_zero_to_one}
\title{Scale a vector to be between zero and one}
\usage{
scale_zero_to_one(x, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{...}{Additional arguments passed to `min` and `max`. Most commonly,
`na.rm`.}
}
\value{
`x` transformed to range between zero and one.
}
\description{
Given a numeric vector returns the vector scaled so it ranges between zero
and one.
}
