% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stat_mode}
\alias{stat_mode}
\title{Find the mode of a vector}
\usage{
stat_mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{A logical value indicating whether `NA` values should be
stripped before the computation proceeds.}
}
\value{
A unit numeric vector representing the most frequent value in `x`
}
\description{
Given a numeric vector returns the most frequently occurring value.
}
