% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_extract_group}
\alias{str_extract_group}
\title{Extract a single match group}
\usage{
str_extract_group(string, pattern, group = 1)
}
\arguments{
\item{group}{Which capture group to extract. Integer.}
}
\description{
Extract a single match group
}
\details{
Meant to make the common pattern of `str_match()[, 2]` easier to
  write and easier to read.
}
