% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{submit_query}
\alias{submit_query}
\title{Submit a query over a connection}
\usage{
submit_query(query, con, params = NULL)
}
\arguments{
\item{query}{An atomic character vector giving either a raw SQL query or
the file path to a .sql file}

\item{con}{A database connection object usable by the `DBI` package}

\item{params}{A named list of parameters to pass to the query. Use "{param_name}"
in `query` to indicate a parameter.}
}
\value{
A data.frame with the result of running `query`
}
\description{
Submit a query over a connection
}
