% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stat_mode}
\alias{stat_mode}
\title{Find the mode of a vector}
\usage{
stat_mode(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A unit numeric vector representing the most frequent value in \code{x}
}
\description{
Given a numeric vector returns the most frequently occurring value.
}
